v0.1.65: Fix NumberFormatException in Position.timestamp and improve error display

+++ app/src/main/java/com/owlsoda/pageportal/services/storyteller/StorytellerModels.kt
@@ -135,7 +135,7 @@ data class ReadAloudResponse(
 
 data class Position(
     val locator: Locator,
-    val timestamp: Long = System.currentTimeMillis()
+    val timestamp: Double = System.currentTimeMillis().toDouble()
 )
 
 data class Locator(
